added segments and rays, the last tools!
still to be done: rework intersection tool to fit segments and rays, fix bug in touchesEnded on measureRatio-Product-Sum-Difference that won't allow one to be created if a different one was already created.

make segments, rays, lines, and circles easier to add: touchBegin, touchMove, and touchEnd to make start and end points and add whatever.